Change dtypes in attc arrays
in search_attc function, when tables are split into several tables if attC sites are distant of more than 4kb, then, the output tables are created using a list comprehension. By default, all columns of those new dataframes are consdered as 'object'. We now convert the position columns to 'int', and the evalue to 'float', to keep consistent with the types, and the frame given by read_infernal.