:bug: change chunk name to avoid inputfile over writing
when the replicon file contains only one sequence
and the sequence id is the same as the file name (without  the
extention) the chunk produce was postfix with nuber between parenthesis
but nextflow generate a shell script with this file name without
escaping the parenthesis.
SO the the script failed.